+++ src/lib.rs
@@ -95,6 +95,8 @@ mod cenum;
 #[cfg(feature = "khr_display_swapchain_9")] pub use khr_display_swapchain::*;
 #[cfg(feature = "khr_get_physical_device_properties2_1")] mod khr_get_physical_device_properties2;
 #[cfg(feature = "khr_get_physical_device_properties2_1")] pub use khr_get_physical_device_properties2::*;
+#[cfg(feature = "khr_get_surface_capabilities2_1")] mod khr_get_surface_capabilities2;
+#[cfg(feature = "khr_get_surface_capabilities2_1")] pub use khr_get_surface_capabilities2::*;
 #[cfg(feature = "khr_incremental_present_1")] mod khr_incremental_present;
 #[cfg(feature = "khr_incremental_present_1")] pub use khr_incremental_present::*;
 #[cfg(feature = "khr_maintenance1_1")] mod khr_maintenance1;