+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current Git Master Branch
 
  * Vulkan 1.0.49
+ * Add extension `VK_KHR_get_surface_capabilities2` (revision 1)
 
 ## Version 0.13.1, released on 07.05.2017